GameConn packet에서 채팅서버 정보 제거 및 관련 코드 수정. game server에서 config 값 활용 없이 listen port 설정하던 부분 수정. 그 외 각 config class 코드 개선 등등

+++ Server/ConfigClass/MasterConfig.h
@@ -7,7 +7,7 @@ public:
 	MasterConfig(const MasterConfig& _other) = delete;
 public:
 	string hostStr = "NONE";
-	int32_t port = 0;
+	uint16_t port = 0;
 public:
 	void Init(rapidjson::Value& _doc);
 	void Render();
